catalog_service/requirements.txt
# catalog_service/requirements.txt

uvicorn
fastapi
python-dotenv
SQLAlchemy
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic-settings
python-jose
psycopg2-binary
asyncpg
slowapi
alembic
httpx
redis